Refactor code and indent

+++ contracts/SolarswapFactory.sol
@@ -4,7 +4,8 @@ import "./interfaces/ISolarswapFactory.sol";
 import "./SolarswapPair.sol";
 
 contract SolarswapFactory is ISolarswapFactory {
-    bytes32 public constant INIT_CODE_PAIR_HASH = keccak256(abi.encodePacked(type(SolarswapPair).creationCode));
+    bytes32 public constant INIT_CODE_PAIR_HASH =
+        keccak256(abi.encodePacked(type(SolarswapPair).creationCode));
 
     address public feeTo;
     address public feeToSetter;
@@ -12,21 +13,34 @@ contract SolarswapFactory is ISolarswapFactory {
     mapping(address => mapping(address => address)) public getPair;
     address[] public allPairs;
 
-    event PairCreated(address indexed token0, address indexed token1, address pair, uint);
+    event PairCreated(
+        address indexed token0,
+        address indexed token1,
+        address pair,
+        uint256
+    );
 
     constructor(address _feeToSetter) public {
         feeToSetter = _feeToSetter;
     }
 
-    function allPairsLength() external view returns (uint) {
+    function allPairsLength() external view returns (uint256) {
         return allPairs.length;
     }
 
-    function createPair(address tokenA, address tokenB) external returns (address pair) {
-        require(tokenA != tokenB, 'Solarswap: IDENTICAL_ADDRESSES');
-        (address token0, address token1) = tokenA < tokenB ? (tokenA, tokenB) : (tokenB, tokenA);
-        require(token0 != address(0), 'Solarswap: ZERO_ADDRESS');
-        require(getPair[token0][token1] == address(0), 'Solarswap: PAIR_EXISTS'); // single check is sufficient
+    function createPair(address tokenA, address tokenB)
+        external
+        returns (address pair)
+    {
+        require(tokenA != tokenB, "Solarswap: IDENTICAL_ADDRESSES");
+        (address token0, address token1) = tokenA < tokenB
+            ? (tokenA, tokenB)
+            : (tokenB, tokenA);
+        require(token0 != address(0), "Solarswap: ZERO_ADDRESS");
+        require(
+            getPair[token0][token1] == address(0),
+            "Solarswap: PAIR_EXISTS"
+        ); // single check is sufficient
         bytes memory bytecode = type(SolarswapPair).creationCode;
         bytes32 salt = keccak256(abi.encodePacked(token0, token1));
         assembly {
@@ -40,12 +54,12 @@ contract SolarswapFactory is ISolarswapFactory {
     }
 
     function setFeeTo(address _feeTo) external {
-        require(msg.sender == feeToSetter, 'Solarswap: FORBIDDEN');
+        require(msg.sender == feeToSetter, "Solarswap: FORBIDDEN");
         feeTo = _feeTo;
     }
 
     function setFeeToSetter(address _feeToSetter) external {
-        require(msg.sender == feeToSetter, 'Solarswap: FORBIDDEN');
+        require(msg.sender == feeToSetter, "Solarswap: FORBIDDEN");
         feeToSetter = _feeToSetter;
     }
 }
